+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 **Enhancements:**
 
-* Added ability to rate page in comments as guest or logged in user, #156
-* Added ability to edit comment or reply to a comment right above it, #164
+* Added ability to rate page in comments as guest or logged in user (possible to disable from admin), #156
+* Added ability to edit comment or reply to a comment right below it, #164
+* Added new option to enable/disable "Read more" for long comment, #169
 
 **Fixes:**